[refactor/#117] 상태 별 책 목록 조회 API 수정

+++ src/main/java/com/techeer/checkIt/domain/review/dto/request/CreateReviewReq.java
@@ -17,5 +17,5 @@ public class CreateReviewReq {
   private long bookId;
   private String title;
   private String content;
-  private Double grade;
+  private int grade;
 }
