+++ src/main/java/com/techeer/checkIt/domain/review/dto/response/ReviewRes.java
@@ -14,5 +14,5 @@ import lombok.Getter;
 public class ReviewRes {
   private String title;
   private String content;
-  private Double grade;
+  private int grade;
 }
